fix: extract shared authenticated frontend fetch helper
