+++ apps/frontend/src/lib/authenticatedFetch.ts
@@ -1,0 +1,32 @@
+const API_BASE_URL = import.meta.env.VITE_API_BASE_URL ?? 'http://localhost:4000';
+
+function getAuthToken(): string | null {
+  return (
+    localStorage.getItem('authToken') ??
+    localStorage.getItem('accessToken') ??
+    localStorage.getItem('token')
+  );
+}
+
+export async function authenticatedFetch<T>(path: string, options: RequestInit = {}): Promise<T> {
+  const token = getAuthToken();
+
+  const response = await fetch(`${API_BASE_URL}${path}`, {
+    ...options,
+    headers: {
+      'Content-Type': 'application/json',
+      ...(token ? { Authorization: `Bearer ${token}` } : {}),
+      ...options.headers,
+    },
+  });
+
+  if (!response.ok) {
+    throw new Error(`Request failed: ${response.status}`);
+  }
+
+  if (response.status === 204) {
+    return undefined as T;
+  }
+
+  return response.json() as Promise<T>;
+}
